+++ scripts/spell-darkness.js
@@ -1,25 +1,28 @@
 /*
-	Magical darkness spreads from a point you choose within range to fill a 15-foot-radius sphere for the duration. The darkness spreads around corners. A creature with darkvision can't see through this darkness, and nonmagical light can't illuminate it.
+	Magical darkness spreads from a point you choose within range to fill a 15-foot-radius sphere for the duration. The
+	darkness spreads around corners. A creature with darkvision can't see through this darkness, and nonmagical light
+	can't illuminate it.
 
-	If the point you choose is on an object you are holding or one that isn't being worn or carried, the darkness emanates from the object and moves with it. Completely covering the source of the darkness with an opaque object, such as a bowl or a helm, blocks the darkness.
+	If the point you choose is on an object you are holding or one that isn't being worn or carried, the darkness
+	emanates from the object and moves with it. Completely covering the source of the darkness with an opaque object,
+	such as a bowl or a helm, blocks the darkness.
 
-	If any of this spell's area overlaps with an area of light created by a spell of 2nd level or lower, the spell that created the light is dispelled.
+	If any of this spell's area overlaps with an area of light created by a spell of 2nd level or lower, the spell that
+	created the light is dispelled.
 */
-const version = "10.2";
+const version = "11.0";
 const optionName = "Darkness";
 
 try {
-	if (args[0].macroPass === "postActiveEffects") {
+	if (args[0].macroPass === "templatePlaced") {
 		// get the template
-		const lastArg = args[args.length - 1];
-
-		let templateDoc = canvas.scene.collections.templates.get(lastArg.templateId);
+		let templateDoc = canvas.scene.collections.templates.get(workflow.templateId);
 		if (templateDoc) {
 			// add the walls to block vision
 			let radius = canvas.grid.size * (templateDoc.distance / canvas.grid.grid.options.dimensions.distance);
 			await circleWall(templateDoc.x, templateDoc.y, radius);
 		}
-	}	
+	}
 	else if (args[0] === "off") {
 		async function removeWalls() {
 			let darkWalls = canvas.walls.placeables.filter(w => w.document.flags["midi-srd"]?.Darkness?.ActorId === actor.id)