+++ scripts/spell-fog-cloud.js
@@ -1,25 +1,22 @@
 /*
-	You create a 20-foot-radius sphere of fog centered on a point within range. The sphere spreads around corners, and its area is heavily obscured. It lasts for the duration or until a wind of moderate or greater speed (at least 10 miles per hour) disperses it.
+	You create a 20-foot-radius sphere of fog centered on a point within range. The sphere spreads around corners, and
+	its area is heavily obscured. It lasts for the duration or until a wind of moderate or greater speed (at least 10
+	miles per hour) disperses it.
 
-	At Higher Levels. When you cast this spell using a spell slot of 2nd level or higher, the radius of the fog increases by 20 feet for each slot level above 1st.
+	At Higher Levels. When you cast this spell using a spell slot of 2nd level or higher, the radius of the fog
+	increases by 20 feet for each slot level above 1st.
 */
-const version = "10.0.0";
+const version = "11.0";
 const optionName = "Fog Cloud";
 
 try {
-	if (args[0].macroPass === "postActiveEffects") {
+	if (args[0].macroPass === "templatePlaced") {
 		// get the template
-		const lastArg = args[args.length - 1];
-		const actorToken = canvas.tokens.get(lastArg.tokenId);
-		
-		let templateDoc = canvas.scene.collections.templates.get(lastArg.templateId);
+		let templateDoc = canvas.scene.collections.templates.get(workflow.templateId);
 		if (templateDoc) {
 			// add the walls to block vision
 			let radius = canvas.grid.size * (templateDoc.distance / canvas.grid.grid.options.dimensions.distance);
 			await circleWall(templateDoc.x, templateDoc.y, radius);
-			
-			// delete the template
-			await canvas.scene.deleteEmbeddedDocuments("MeasuredTemplate", [templateDoc.id]);
 		}
 	}	
 	else if (args[0] === "off") {