+++ scripts/spell-healing-spirit.js
@@ -1,38 +1,42 @@
 /*
-	You call forth a nature spirit to soothe the wounded. The intangible spirit appears in a space that is a 5-foot cube you can see within range. The spirit looks like a transparent beast or fey (your choice).
+	You call forth a nature spirit to soothe the wounded. The intangible spirit appears in a space that is a 5-foot cube
+	you can see within range. The spirit looks like a transparent beast or fey (your choice).
 
-	Until the spell ends, whenever you or a creature you can see moves into the spirit’s space for the first time on a turn or starts its turn there, you can cause the spirit to restore 1d6 hit points to that creature (no action required). The spirit can’t heal constructs or undead. The spirit can heal a number of times equal to 1 + your spellcasting ability modifier (minimum of twice). After healing that number of times, the spirit disappears.
+	Until the spell ends, whenever you or a creature you can see moves into the spirit’s space for the first time on a
+	turn or starts its turn there, you can cause the spirit to restore 1d6 hit points to that creature (no action required).
+	The spirit can’t heal constructs or undead. The spirit can heal a number of times equal to 1 + your spellcasting
+	ability modifier (minimum of twice). After healing that number of times, the spirit disappears.
 
 	As a bonus action on your turn, you can move the spirit up to 30 feet to a space you can see.
 
-	At Higher Levels. When you cast this spell using a spell slot of 3rd level or higher, the healing increases by 1d6 for each slot level above 2nd.
+	At Higher Levels. When you cast this spell using a spell slot of 3rd level or higher, the healing increases by 1d6
+	for each slot level above 2nd.
 */
-const version = "10.3";
+const version = "11.0";
 const optionName = "Healing Spirit";
-const actorName = "Healing Spirit";
 const summonFlag = "healing-spirit";
 
 try {
 	const lastArg = args[args.length - 1];
-	let caster = MidiQOL.MQfromActorUuid(lastArg.actorUuid);
-	const actorToken = canvas.tokens.get(lastArg.tokenId);
-	
+
     if (args[0] === "on") {
-		const sourceItem = await fromUuid(lastArg.origin);
-		let summonName = optionName + " (" + caster.name + ")";
-		//const spellLevel = Number(args[1]);
+		// collect cast data
 		const spellLevel = lastArg.efData.flags["midi-qol"].castData.castLevel;
-		const ability = caster.system.attributes.spellcasting;
-		const abilityBonus = caster.system.abilities[ability].mod;
-		const uses = Math.max(abilityBonus + 1, 2);
 		const healDice = spellLevel - 1;
+
+		const ability = actor.system.attributes.spellcasting;
+		const abilityBonus = actor.system.abilities[ability].mod;
+		const uses = Math.max(abilityBonus + 1, 2);
+
+		// bonus for certain features
 		let healingBonus = 0;
 
-		let featureItem = caster.items.getName("Disciple of Life");
+		let featureItem = actor.items.getName("Disciple of Life");
 		if (featureItem) {
 			healingBonus = 2 + spellLevel;
-		}		
+		}
 
+		const summonName = optionName + " (" + actor.name + ")";
         let updates = {
             token: {
 				"name": summonName,
@@ -41,7 +45,7 @@ try {
 				"displayBars": CONST.TOKEN_DISPLAY_MODES.ALWAYS,
 				"bar1": { attribute: "attributes.hp" },
 				"actorLink": false,
-				"flags": { "midi-srd": { "Healing Spirit": { "ActorId": caster.id } } }
+				"flags": { "midi-srd": { "Healing Spirit": { "ActorId": actor.id } } }
             },
 			"name": summonName,
 			embedded: {
@@ -76,8 +80,8 @@ try {
 		}
 
 		// spawn the actor
-		const maxRange = sourceItem.system.range.value ? sourceItem.system.range.value : 60;
-		let position = await HomebrewMacros.warpgateCrosshairs(actorToken, maxRange, sourceItem, summonActor.prototypeToken);
+		const maxRange = item.system.range.value ? item.system.range.value : 60;
+		let position = await HomebrewMacros.warpgateCrosshairs(token, maxRange, item);
 		if (position) {
 			// check for token collision
 			const newCenter = canvas.grid.getSnappedPosition(position.x - summonActor.prototypeToken.width / 2, position.y - summonActor.prototypeToken.height / 2, 1);
@@ -86,7 +90,7 @@ try {
 				return false;
 			}
 
-			const result = await warpgate.spawnAt(position, summonName, updates, { controllingActor: caster }, {});
+			const result = await warpgate.spawnAt(position, summonName, updates, { controllingActor: actor }, {});
 			if (!result || !result[0]) {
 				ui.notifications.error(`${optionName} - Unable to spawn`);
 				return false;
@@ -94,10 +98,7 @@ try {
 
 			let summonedToken = canvas.tokens.get(result[0]);
 			if (summonedToken) {
-				await caster.setFlag("midi-qol", summonFlag, summonedToken.id);
-				// players can't do the following:
-				//await summonedToken.toggleCombat();
-				//await summonedToken.actor.rollInitiative();
+				await actor.setFlag("midi-qol", summonFlag, summonedToken.id);
 			}
 			
 		}
@@ -109,9 +110,9 @@ try {
 	}
 	else if (args[0] === "off") {
 		// delete the summon
-		const lastSummon = caster.getFlag("midi-qol", summonFlag);
+		const lastSummon = actor.getFlag("midi-qol", summonFlag);
 		if (lastSummon) {
-			await caster.unsetFlag("midi-qol", summonFlag);
+			await actor.unsetFlag("midi-qol", summonFlag);
 			await warpgate.dismiss(lastSummon, game.canvas.scene.id);
 		}
 	}