+++ scripts/spell-wall-of-sand.js
@@ -1,30 +1,28 @@
 /*
 	You conjure up a wall of swirling sand on the ground at a point you can see within range. You can make the wall up to 30 feet long, 10 feet high, and 10 feet thick, and it vanishes when the spell ends. It blocks line of sight but not movement. A creature is Blinded while in the wall’s space and must spend 3 feet of movement for every 1 foot it moves there.
 */
-const version = "10.0.0";
+const version = "11.0";
 const optionName = "Wall of Sand";
 const templateFlag = "wall-of-sand-template";
 
 try {
-	const lastArg = args[args.length - 1];
+	if (args[0].macroPass === "templatePlaced") {
+		// get the template
+		let templateDoc = canvas.scene.collections.templates.get(workflow.templateId);
+		if (templateDoc) {
+			// add the walls to block vision
+			await lineWall(templateDoc);
 
-	if (args[0].macroPass === "postActiveEffects") {
-		// build the template macro
-		let templateDoc = canvas.scene.collections.templates.get(lastArg.templateId);
-		if (!templateDoc) return;
+			// store the spell data in the template
+			let touchedTokens = await game.modules.get('templatemacro').api.findContained(templateDoc);
+			await templateDoc.setFlag('world', 'spell.WallofSand', {touchedTokens});
 
-		// add the walls to block vision
-		await lineWall(templateDoc);
-
-		// store the spell data in the template
-		let touchedTokens = await game.modules.get('templatemacro').api.findContained(templateDoc);
-		await templateDoc.setFlag('world', 'spell.WallofSand', {touchedTokens});
-		
-		// apply blinded to the tokens
-		for (let token of touchedTokens) {
-			let tokenDoc = canvas.scene.tokens.get(token);
-			await game.dfreds?.effectInterface.addEffect(
-			{ effectName: 'Blinded', uuid: tokenDoc.actor.uuid, origin: lastArg.templateId });
+			// apply blinded to the tokens
+			for (let token of touchedTokens) {
+				let tokenDoc = canvas.scene.tokens.get(token);
+				await game.dfreds?.effectInterface.addEffect(
+					{ effectName: 'Blinded', uuid: tokenDoc.actor.uuid, origin: workflow.templateId });
+			}
 		}
 	}
 	else if (args[0] === "off") {